refactor: rename oauth schema module and update imports
