+++ registry/src/registry/auth/oauth/oauth_client.py
@@ -8,7 +8,7 @@ import httpx
 from authlib.integrations.httpx_client import AsyncOAuth2Client
 from authlib.oauth2.rfc7636 import create_s256_code_challenge
 
-from registry.schemas.oauth_sachema import (
+from registry.schemas.oauth_schema import (
     MCPOAuthFlowMetadata,
     OAuthClientInformation,
     OAuthMetadata,
