+++ registry/src/registry/auth/oauth/redis_flow_storage.py
@@ -6,7 +6,7 @@ from redis import Redis
 
 from registry.constants import REGISTRY_CONSTANTS
 from registry.schemas.enums import OAuthFlowStatus
-from registry.schemas.oauth_sachema import MCPOAuthFlowMetadata, OAuthFlow, OAuthTokens
+from registry.schemas.oauth_schema import MCPOAuthFlowMetadata, OAuthFlow, OAuthTokens
 
 logger = logging.getLogger(__name__)
