+++ registry/src/registry/schemas/oauth_schema.py
