+++ registry/tests/unit/services/test_token_service.py
@@ -8,7 +8,7 @@ import pytest
 from beanie import PydanticObjectId
 
 from registry.schemas.enums import TokenType
-from registry.schemas.oauth_sachema import OAuthClientInformation, OAuthTokens
+from registry.schemas.oauth_schema import OAuthClientInformation, OAuthTokens
 from registry.services.oauth.token_service import TokenService
 from registry_pkgs.models._generated.token import Token
